fix(sitl): cast blackbox log number for %ld in OSD stats (64-bit hosts)
The blackbox log-number stat compiles on SITL only once USE_SDCARD is
enabled (added for SITL SD-card support in this branch). On 64-bit hosts
int32_t is not long, and clang on macOS errors on the format mismatch
with warnings-as-errors. Cast the argument to match the format.

+++ src/main/io/osd.c
@@ -5459,7 +5459,7 @@ static void osdShowStats(bool isSinglePageStatsCompatible, uint8_t page)
 
                 int32_t logNumber = blackboxGetLogNumber();
                 if (logNumber >= 0) {
-                    tfp_sprintf(buff, ": %05ld ", logNumber);
+                    tfp_sprintf(buff, ": %05ld ", (long)logNumber);
                 } else {
                     strcat(buff, ": INVALID");
                 }
